updated dataset_processor test
